added theme to index file
